feat: UIScrollView StopScrolling category renamed to TWScrolling, new methods implemented
